usb: gadget: Do not drop the request upon suspend
If a write(IN) request is queued after usb bus suspend in smd driver,
it queues the request in tx_pull which fails with -EAGAIN. As a result
response/indication generating a remote wakeup signal is dropped
and never received by the host.

Fix this by not dropping response packets on enqueue error
but now wake up to the host connected and queue the request on resume.

Change-Id: I2272b728699159e14581d8415b6ad9617d95685f
Signed-off-by: ChandanaKishori Chiluveru <cchilu@codeaurora.org>