net: danipc: move M-FIFO buffers to B-FIFO
In current code on device open, new buffers are added to B-FIFO.  When
device is opened a second time, these buffers could already exist in the
M-FIFO or could be in transit.  This results in duplicate buffers in the
B-FIFO, which results in a race condition.  Therefore, move M-FIFO buffers
to B-FIFO when the device is opened the second time instead of adding new
buffers to B-FIFO.

Change-Id: Idc1ec73e7d54fb2ea76d092bb944278868c3217d
Acked-by: Yasir Malik <ymalik@qti.qualcomm.com>
Signed-off-by: Sivanesan Rajapupathi <srajap@codeaurora.org>