msm: sps: driver performance improvement
sps driver can be accessed heavily in high throughput use case
such as ipsec with crypto driver as client. The following changes are made
to improve driver performance.

CPU reading BAM registers is slow. The write descriptor
offset information in the BAM_P_EVENT_REG is only set by the host.
Therefore, the write descriptor offset information is cached in the driver
during write. Subsequent read will read the cached information
in the driver, instead of going out to read from BAM for the performance.

Furthermore, the BAM descriptor read and write pointers are accessed
heavily in data transfer. For each transfer the register offset is
calculated.  The register offset information is static and never changed.
Therefore, for the performance reason, we calculate it once in sps_connect
and store it aside. Subsequent data transfer uses the previous result
without doing another calculation.

In the functions involved in data transfer, the sps_pipe structure
pointer is passing in as an argument, instead of passing in pipe index
to save de-referencing pointer calculation in every invocation of function.

Add global flag in sps driver to by-pass error checking in sps_transfer,
if the flag is set. This flag should only be set after client code above
driver is well debugged. The overhead of error checking may be
significant under heavy traffic. Later, we may add API to enable/disable
this flag from client.

Change-Id: If3adf44368e862c25ede9bb061b5e5507261091f
Acked-by: Che-Min Hsieh <cheminh@qti.qualcomm.com>
Signed-off-by: Sivanesan Rajapupathi <srajap@codeaurora.org>