net: danipc: Implement the mmap on chardev interface
To improve the performance, the mmap is implemented on the character
device interface to avoid the data copy. The buffer management is
implemented in the driver to track the TX/RX buffers. The new ioctl
commands are introduced for buffer management, packet transmission
and packet receiving in the mmapped address space.

Change-Id: Id170773eea0535ef388b053714eb316d69db9fc1
Acked-by: Jie Luo <jluo@qti.qualcomm.com>
Signed-off-by: Sivanesan Rajapupathi <srajap@codeaurora.org>