msm: camera: add support for buffer drop
If frame request for frame N comes after SOF N, buffer is dropped. We have
used this feature to drop frames intentionally from user space. But
this changes buffer order, because in the design is assumed that buffer
will be dequeued consequently. Because of that frame drop option is added
to frame request ioctl. Now we can drop buffer before start of FE.

Change-Id: I03d8513a1fe364fe3254de554928d56251d0f75a
Signed-off-by: Konstantin Motov <kmotov@codeaurora.org>